[refactor]#108 유저 게시글 추천  interface 분리
