+++ src/main/java/org/netway/dongnehankki/recommendation/presentation/RecommendationController.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.netway.dongnehankki.global.auth.CustomUserDetails;
 import org.netway.dongnehankki.global.common.ApiResponse;
 import org.netway.dongnehankki.recommendation.application.RecommendationService;
+import org.netway.dongnehankki.recommendation.application.RecommendationServiceImpl;
 import org.netway.dongnehankki.post.domain.Post;
 import org.netway.dongnehankki.post.dto.response.PostResponse;
 import org.springframework.http.ResponseEntity;
@@ -33,12 +34,10 @@ public class RecommendationController {
         @AuthenticationPrincipal CustomUserDetails userDetails,
         @Parameter(description = "가져올 게시글 수") @RequestParam(defaultValue = "10") int limit) {
 
-        Long userId = userDetails.getUser().getUserId(); // 로그인한 사용자 ID
+        Long userId = userDetails.getUser().getUserId();
 
         List<Post> recommendedPosts = recommendationService.getRecommendedPosts(userId, limit);
 
-        // Post 엔티티를 PostResponse DTO로 변환
-        // 추천 게시글의 isLiked는 기본적으로 false로 설정 (사용자가 아직 좋아요를 누르지 않은 게시글이므로)
         List<PostResponse> response = recommendedPosts.stream()
             .map(post -> PostResponse.fromEntity(post))
             .collect(Collectors.toList());
